# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/coach/access": {
      "get": {
        "operationId": "AiChatController_getAccess",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ai"
        ]
      }
    },
    "/v1/coach/chat": {
      "post": {
        "operationId": "AiChatController_reply",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiChatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ai"
        ]
      }
    },
    "/v1/coach/chat/stream": {
      "post": {
        "operationId": "AiChatController_replyStream",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AiChatDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ai"
        ]
      }
    },
    "/v1/coach/conversations": {
      "get": {
        "operationId": "AiChatController_listConversations",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ai"
        ]
      }
    },
    "/v1/coach/conversations/{id}/messages": {
      "get": {
        "operationId": "AiChatController_listConversationMessages",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ai"
        ]
      }
    },
    "/v1/coach/conversations/{id}": {
      "delete": {
        "operationId": "AiChatController_deleteConversation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ai"
        ]
      }
    },
    "/v1/coach/messages/{id}/feedback": {
      "patch": {
        "operationId": "AiChatController_setFeedback",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CoachFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ai"
        ]
      }
    },
    "/v1/coach/memory": {
      "get": {
        "operationId": "AiChatController_getMemory",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ai"
        ]
      },
      "delete": {
        "operationId": "AiChatController_clearMemory",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "ai"
        ]
      }
    },
...
  },
  "info": {
    "title": "Mentor API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "AiChatDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "clientMessageId": {
            "type": "string",
            "format": "uuid"
          },
          "conversationId": {
            "type": "string",
            "format": "uuid"
          },
          "contextMockExamId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "message"
        ]
      },
      "CoachFeedbackDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}